+++ src/TestSuite/TableTestCase.php
@@ -50,45 +50,49 @@ abstract class TableTestCase extends TestCase
     /**
      * Asserts that the table has a "belongs to" association
      * @param \Cake\ORM\Association $association Association
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
      */
-    public function assertBelongsTo(Association $association): void
+    public function assertBelongsTo(Association $association, string $message = ''): void
     {
-        $this->assertInstanceOf(BelongsTo::class, $association);
+        $message = $message ?: sprintf('Failed asserting that `%s` is an instance of `%s`', get_class($association), BelongsTo::class);
+        $this->assertInstanceOf(BelongsTo::class, $association, $message);
     }
 
     /**
      * Asserts that the table has a "belongs to many" association
      * @param \Cake\ORM\Association $association Association
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
-     * @todo missing params
      */
-    public function assertBelongsToMany(Association $association): void
+    public function assertBelongsToMany(Association $association, string $message = ''): void
     {
-        $this->assertInstanceOf(BelongsToMany::class, $association);
+        $message = $message ?: sprintf('Failed asserting that `%s` is an instance of `%s`', get_class($association), BelongsToMany::class);
+        $this->assertInstanceOf(BelongsToMany::class, $association, $message);
     }
 
     /**
      * Asserts that the table has a behavior
-     * @param string|array $behavior Behavior name as string or array
+     * @param string|string[] $behavior Behavior name as string or array
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
-     * @todo missing params
      */
-    public function assertHasBehavior($behavior): void
+    public function assertHasBehavior($behavior, string $message = ''): void
     {
         foreach ((array)$behavior as $name) {
-            $this->assertTrue($this->Table->hasBehavior($name));
+            $this->assertTrue($this->Table->hasBehavior($name), $message ?: sprintf('Failed asserting that `%s` has `%s` behavior', $this->originClassName, $name));
         }
     }
 
     /**
      * Asserts that the table has a "many" association
      * @param \Cake\ORM\Association $association Association
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
-     * @todo missing params
      */
-    public function assertHasMany(Association $association): void
+    public function assertHasMany(Association $association, string $message = ''): void
     {
-        $this->assertInstanceOf(HasMany::class, $association);
+        $message = $message ?: sprintf('Failed asserting that `%s` is an instance of `%s`', get_class($association), HasMany::class);
+        $this->assertInstanceOf(HasMany::class, $association, $message);
     }
 }