+++ tests/TestCase/TestSuite/TableTestCaseTest.php
@@ -15,8 +15,11 @@ declare(strict_types=1);
 
 namespace MeCms\Test\TestCase\TestSuite;
 
+use MeCms\Model\Table\PostsCategoriesTable;
 use MeCms\Model\Table\PostsTable;
+use MeCms\Test\TestCase\Model\Entity\UserTest;
 use MeCms\Test\TestCase\Model\Table\PostsTableTest;
+use MeCms\Test\TestCase\Model\Table\UsersTableTest;
 use MeCms\TestSuite\TestCase;
 
 /**
@@ -34,5 +37,58 @@ class TableTestCaseTest extends TestCase
         $this->assertSame('Posts', $TableTestCase->alias);
         $this->assertSame(PostsTable::class, $TableTestCase->originClassName);
         $this->assertInstanceOf(PostsTable::class, $TableTestCase->Table);
+    }
+
+    /**
+     * @test
+     * @uses \MeCms\TestSuite\TableTestCase::assertBelongsTo()
+     */
+    public function testAssertBelongsTo(): void
+    {
+        $TableTestCase = new PostsTableTest();
+        $TableTestCase->assertBelongsTo($TableTestCase->Table->Categories);
+
+        $this->expectAssertionFailed('Failed asserting that `Cake\ORM\Association\BelongsToMany` is an instance of `Cake\ORM\Association\BelongsTo`');
+        $TableTestCase->assertBelongsTo($TableTestCase->Table->Tags);
+    }
+
+    /**
+     * @test
+     * @uses \MeCms\TestSuite\TableTestCase::assertBelongsToMany()
+     */
+    public function testAssertBelongsToMany(): void
+    {
+        $TableTestCase = new PostsTableTest();
+        $TableTestCase->assertBelongsToMany($TableTestCase->Table->Tags);
+
+        $this->expectAssertionFailed('Failed asserting that `Cake\ORM\Association\BelongsTo` is an instance of `Cake\ORM\Association\BelongsToMany`');
+        $TableTestCase->assertBelongsToMany($TableTestCase->Table->Categories);
+    }
+
+    /**
+     * @test
+     * @uses \MeCms\TestSuite\TableTestCase::assertHasBehavior()
+     */
+    public function testAssertHasBehavior(): void
+    {
+        $TableTestCase = new PostsTableTest();
+        $TableTestCase->assertHasBehavior('Timestamp');
+        $TableTestCase->assertHasBehavior(['CounterCache', 'Timestamp']);
+
+        $this->expectAssertionFailed('Failed asserting that `MeCms\Model\Table\PostsTable` has `NoExistingBehavior` behavior');
+        $TableTestCase->assertHasBehavior(['CounterCache', 'NoExistingBehavior']);
+    }
+
+    /**
+     * @test
+     * @uses \MeCms\TestSuite\TableTestCase::assertHasMany()
+     */
+    public function testAssertHasMany(): void
+    {
+        $TableTestCase = new UsersTableTest();
+        $TableTestCase->assertHasMany($TableTestCase->Table->Posts);
+
+        $this->expectAssertionFailed('Failed asserting that `Cake\ORM\Association\BelongsTo` is an instance of `Cake\ORM\Association\HasMany`');
+        $TableTestCase->assertHasMany($TableTestCase->Table->Groups);
     }
 }